added types in fype file
