+++ src/type/Todo.tsx
@@ -1,0 +1,9 @@
+import { User } from './User';
+
+export interface Todo {
+  userId: number;
+  id: number;
+  title: string;
+  completed: boolean;
+  user: User | undefined;
+}
